refactor(core): Remove accidental cross-module imports
- oauth-server no longer imports INSTANCE_MCP_RESOURCE_ID from the mcp
  module; the constant now lives in oauth-server and mcp imports it,
  matching the existing init-time dependency direction.
- breaking-changes' unverified-packages rule gets its own minimal config
  mirroring CommunityPackagesConfig's effective enabled flag instead of
  importing the community-packages module's config.
- resolveBuiltinNodeDefinitionDirs moves from the instance-ai module to
  a shared cli util, removing the agents -> instance-ai import.

Part of DEVP-646.
